Move i18n configuration module to `configs` folder
