+++ apps/user-portal/src/configs/index.ts
@@ -17,5 +17,6 @@
  */
 
 export * from "./app";
+export * from "./i18n";
 export * from "./routes";
 export * from "./ui";
